+++ resources/views/layouts/navi-header.blade.php
@@ -141,7 +141,7 @@ If not, see <https://www.gnu.org/licenses/>. -->
                 </span>
             </button>
 
-            @if(!in_array(Route::currentRouteName(), ['translate', 'log.detail']))
+            @if(!in_array(Route::currentRouteName(), ['translate']))
                 <div x-data="{ open: false }" class="relative">
                     <button
                         type="button"